requirements.txt
tensorflow
pillow
streamlit